refactor: mover pagina de lugar
se movió el archivo page.tsx de /tickets/assignment a user/assign-ticket para poder usar el middleware de autenticación (sólo users pueden acceder a la página)
